환불기능 임시 구현

+++ src/main/java/com/shop/mapper/user/ReviewMapper.java
@@ -9,9 +9,9 @@ import com.shop.dto.user.review.MyReviewResponseDTO;
 public interface ReviewMapper {
 	void insertReview(Review review);
 
-	List<Review> getReviewListByProduct(int productNo);
+	List<Review> getReviewListByProduct(Long productNo);
 
-	Review getOneReviewByOrderItem(int orderItemNo);
+	Review getOneReviewByOrderItem(Long orderItemNo);
 
 	List<MyReviewResponseDTO> selectReviewsByMemberNo(Long memberNo);
 }
